fix clipping rect float to int conversion
When truncating x/y, the truncated amount needs to be added to w/h
before ceil is applied.  This ensures that the clipping rectangle is
always at least as large as we asked for (in floats).

Previously the low edges of some glyphs (like 'g') got cut off at
certain zoom levels.