fix blurry text/icons on non-hdpi displays
Add "snap_to_pixels" globar var (default true) that controls if text and
icon positions are rounded to nearest pixel.

This helps prevent blurry text whenever the text is centered or zoomed.

You generally want this enabled, unless you are trying to slowly and
smoothly move text on the screen.  (Example is text annotation floating
next to a smoothly moving picture.)

demo() updated to control this and also zoom in/out will scale to font
sizes instead of arbitrarily