fixes to support buttons with given rect
This is a start on modifying widgets to use a local coordinate system
with the origin being their content rect topleft corner.

The default processEvents() uses the border rect for catching mouse
events.  That way clicking on a button's border or margin still works.

But any math on mouse events should use the content rect.