allow adding a dialog from any thread
This is very useful for warnings and error.  In podcast, a
background thread could get an error while accessing the database or
downloading a file.  This makes it easy to show something to the user
without having to manually pass data back to the GUI thread.

Moved dataGet/dataSet into Window with a mutex for all that data.

Separate mutex for dialogs, so you can add a dialog and dataSet all its
data without it being displayed while you are adding data.