textLayout: only set min width if we aren't breaking lines
Fixes the problem where the vertical scrollbar wouldn't get enough width
when the window was being resized to have smaller total width.

TextLayoutWidget doesn't have a normal min width or height (when
breaking lines).  Given a width it will calculate it's min height.  This
change makes it not set it's min width (when breaking lines).