menu: allow mousing over items if towards a submenu
If the mouse goes over a menu item it normally focuses that menu item.
But now if the mouse motion is towards an open submenu popup, then the
menu item won't focus.

This makes it easier for the user to move the mouse to the submenu
without accidentally focusing a different menu item.

Current test is if the mouse motion is within pi/3 (60 deg) of moving
towards the center of the submenu popup.