warn if widget expand is same as containing scroll auto
A scroll area with scrolling .auto in a direction gets it's virtual size
from the contained widgets.  If a contained widget tries to expand in
that direction, we ignore it and log a warning.

Unclear what it would mean to expand in a direction you can scroll in.
It's possible in the future to add support for the situation where a
contained widget could expand to fill as much space as the scroll area
actually has in the viewport.

The motivating example would be a window where you want the final widget
to expand to fill the window, but if the window is too small you want a
scrollbar to show up.