textLayout: assume large width if given rect width is 0
This helps prevent bad flickering on the second to third frames for a
TextLayoutWidget.  The first frame we get rect width 0, so show nothing,
but have to calculate a min size.  The min height depends on the given
width, which we don't know that frame.

We used to pick a small width, which would cause a large min height for
the second frame so you'd see a large mostly-empty TextLayoutWidget on
frame 2, then frame 3 would look okay.

Now we pick a large width, which causes a small min height for frame 2,
so you see a single line of text on frame 2, then frame 3 looks okay.
This seems to look better.