textLayout: increase default min width to 250
This combined with the previous revert gives better behavior when initially
showing a TextLayoutWidget. On the first frame the size is zero, but we don't
know what width to use to calculate our min height for the second frame.

Initially I had it pick a small value (the default min width was 25), but then
the min height was calculated too large, which looked like a frame of flicker.

Then I switched to a large value (arbitrarily 1000), but then the min height
was too small.  That messed up the size calculations for dialogs (which usually
use the textLayout min size to decide the dialog size).

So now I've switched back to using the min width, but bumped the default up to
250.  Hopefully this will be a decent default, and at least it can be
overridden if needed.