textEntry and textLayout: make selection robust against text change
The selection is kept from frame to frame, but the text might change out
from under the widget.  Since the selection is kept in bytes, it might
end up in the middle of a utf8 multibyte char, causing an error.

Now we detect when that happens and push the selection start/cursor/end
to the next non-continuation char.