mouse wheel: propogate the scroll if needed
Wheel events are processed by scroll containers after child widgets so
the inner-most scroll container gets the first chance to process it.

If that container is already at the end of it's scroll, then it lets the
event propogate to the next container.

For the user, if you are using the wheel to scroll in a textEntry, when
you can't scroll anymore then the containing window will scroll.