textLayout: stop measuring text early if we can
Normally TextLayoutWidget must process (measure, if not render) all the
text it receives to calculate its min height, even if most of the text
is outside the clipping region.

If we are given a rect, then we don't need to calculate a min height, so
can stop processing text as soon as it runs off the end of the clipping
region.

Example is the list of episodes in a podcast player.  Each episode is
given a rect to render as much of the episode description as it can.