mouse wheel: propogate the scroll event even when changing orientations
It can be a little weird to have the mouse wheel scroll a text entry
horizontally a bit and then suddenly scroll the containing window.  So
initially that didn't happen.

But it's much weirder to have an empty text entry and not be able to
scroll the containing window if the mouse pointer is over the text
entry.

So now we'll propogate the scroll event even for hitting the edge of a
horizontal scroll with the mouse wheel.