textEntry: separate EventMatchOptions from processEvents
This is part of feeling out how to provide the user a straightforward
way to intercept/override events.

The motivating use-case is having a keyboard shortcut.

I initially tried to have the user process the events before the widget,
but that fails in the situation where the widget gets keyboard focus and
the shortcut in the same frame.  Because when the user processes events,
the widget hasn't gotten focus yet.

So the solution needs to be able to interleave event processing code
from the user and widget.  This is a step in that direction.