better strategy for dataGet/dataSet
dataSet() automatically copies slice contents, because that's usually what
you want.

dataSetAdvanced() can control whether slice contents are copied.

dataGetPtr() returns a pointer to the internal storage, in case you are
storing a large piece of data that's not a slice, and don't want to copy
the data back and forth.