focusWidget: focus the widget in the current subwindow by default
It used to focus the widget in the focused subwindow, but almost always
you are focusing a widget in the context of the current subwindow (which
is the subwindow widgets are being added to).

Now accepts an optional parameter for the rare case you need to focus a
widget in a given subwindow.  Usually this is due to a bubbled event,
like how a ContextWidget gets a close_popup event while the popup is
still the current window.