adjust our frame overhead estimate faster
When you use beginWait()/waitTime() so that dvui manages the framerate
for you, it maintains an estimate of how much time is spent outside of
dvui's begin/end.

This estimate is used as part of what waitTime() returns to more
accurately hit timer deadlines. The estimate is only updated when we are
trying to wait for a timer and are not woken by an event.

Previously this estimate (in microseconds) was adjusted by how many
frames we had been adjusting it in the same direction.

Now we adjust it by 2^number of frames.  This should reduce the amount
of frames it takes for the estimate to find the "true" value.

Also show the internal estimate in the clock example.