normalizing apis
We are moving from passing in options (like whether to process events
and what stuff to draw) to install() to having a few more
functions per widget.

Instead of customizing a widget via options, the
idea is to start with the functions called in the high-level api
(like button()), and then adjust as needed.