matchEvent: directly query widgets for eventMatch
We were using a roundabout method of checking if an event matched
(should be looked at) to a widget.  It broke down when you want to have
a compound widget where the top one gets all the events and directs them
to widgets inside.

Now we call matchEvent() on the widget and it says if the given event
should be looked at.