widgets passively listen to events in matchEvent
This comes up when one widget is wrapping another (like sliderEntry wraps
TextEntry that wraps TextLayout).  In this case, sliderEntry is responsible for
forwarding events to TextEntry (who does the same for TextLayout).

For mouse events this works by sliderEntry asking TextEntry if it might process
this event with `te.matchEvent()`.

For key events this works by sliderEntry having focus, and directly forwarding
the event to TextEntry with `te.processEvent()`.

This does not work for events that TextEntry might passively listen to but
would not mark as handled or return true from `te.matchEvent()`.  Say it wanted
to keep track of whether ctrl was down:
* it can't return true for all key events from `te.matchEvent()` because
sliderEntry uses that to decide which events to intercept
* it can't listen inside `te.processEvent()` because then it would have to do it's own `te.matchEvent()` check inside `te.processEvent()` which would drop key events since te doesn't have focus (sliderEntry does)

So what we are going to try is:
* passively listening for events should happen in `matchEvent()`
* wrapping a widget implies you must call `matchEvent()` for every event, so you can't shortcircuit the logic