colors: trying to cleanup the color_style mess (thanks r4gus)
Previously we were trying to specify colors in Options by style, where each
style was backed by a collection of colors in the Theme.  The idea was
that any widget might easily switch to a different style.

So far that hasn't panned out:
* most widgets only care about the different fill colors
* menu/dropdown items need to be accent styled
* maybe need buttons with accent and err styling?

We are switching to a simpler plan:
* Theme has a flat list of colors
* Theme includes a few premade Options (style_accent, style_err)
* when passing a color into Options, it's now a union of a literal Color
  and a name which means we pick it from the theme