textLayout: better selectionGet api
The old api was that you could get null returned that indicated that the
selection was being changed (so it would be different next frame).  That
caused enough problems that we are trying something else.

The primary way the selection can be invalid is if the start/cursor/end
is beyond the end of the text.  This happens when we do a select all.

So we are trying an API where you have to pass in a max len and always
get a pointer to the selection back, and the start/cursor/end are capped
to that max.