backends: remove free() function
I added it originally to prevent copying the clipboard contents in the
SDLBackend, but it pushes complexity to the caller instead.  Also the
backends are getting too many functions and I'm looking for ways to cut
down.

Now we'll live with the extra copy for a while and see how that goes.