web: able to enter characters into text input
This introduces the idea of the backend showing/hiding the on screen
keyboard.  Currently it's either on or off, but uses a Rect so that in
the future we could support the backend positioning an IME (Input Method
Editor) window appropriately.

On chrome (at least on one phone I tested), the backspace key doesn't
quite work properly.  It first removes the android keyboard autocomplete
suggestions, and only when those are gone does dvui see it.  Need to
investigate further.