text rendering: snap and use smaller font
Testing seems to show that when fonts are scaled (not shown at their
natural size), picking a smaller font and stretching it looks much
better than picking a larger font and squeezing it.  So now we pick a
smaller font.

Streching the font still looks bad when the scale is < 2.

Now streching the font is controlled by Window.snap_to_pixels, which
defaults to true if the initial scale is < 2.