web backend: make on screen keyboard work on safari
Safari restricts when javascript can cause the on screen keyboard to
show.  Javascript can only show it "in the context of a user
interaction", which seems to mean during an event callback.

Now we treat the on screen keyboard state similar to cursor.  Each
frame, after Window.end(), the application must call
Window.OSKRequested() and handle it.

The web-test sample sets a variable in WebBackend.js, which is queried
at the end of mouse up and touch up events.