text layout: fix refresh bug
If the selection end was at the end of a line that was offset by a corner
widget, then it caused a refresh each frame.  This was because the new
sel_end_r is calculated twice, once at the end of that line, and again at the
beginning of the next line.

Both calculations are needed in case the text ends right there, and because we
don't know if the next line will also be offset by the corner widget.

The solution is keep both, but only check for refresh in addTextDone().