frame overhead: start estimate at 1ms
We used to start the estimate at 0, but after some casual observations
of different computers/phones, 1ms is a much better starting place.

This only affects timers (like a clock) that are trying to schedule
frames at certain times.

Most of the time we run frames either in response to input events, or as
fast as we can when running animations (limited by vsync if enabled and
our fps limit passed to waitTime()).