sdl: put back in a bug workaround
https://github.com/libsdl-org/SDL/issues/6539

Before SDL 2.28, there was a bug that would show as a temporary hang
when showing a toast.  DVUI was trying to do a wait with timeout that
could be interrupted by an event, but the bug made it an unconditional
wait until the toast timeout.

We need to keep this workaround in for longer to help people with older
versions of SDL.