Revert "Expand: revert function names from "isVertical" back to "vertical""
This reverts commit 97b151e8da7d8af5803ffbc0c34d40c9322b6944.

This reverts a revert because I didn't understand the problem, which
stems from wanting Theme to be serializable (which includes Option,
which includes Expand).

Serialization requires the fields of Expand to have distinct names, and
in particular to be distinct from functions.  Originally "vertical" was
both a field and a function.

So changing the functions to "isVertical" enables Theme serialization.