Revert "Expand: revert function names from "isVertical" back to "vertical""
This reverts commit e4f9d8939a9066b55f5cc01df9bdf245788ad575.

This reverts a revert because I didn't understand the problem, which
stems from wanting Theme to be serializable (which includes Option,
which includes Expand).

Serialization requires the fields of Expand to have distinct names, and
in particular to be distinct from functions.  Originally "vertical" was
both a field and a function.

So changing the functions to "isVertical" enables Theme serialization.