scrolling: better touch scroll capture and propogation
When a ScrollContainer receives a scroll event that would scroll past
the limit, we want to propogate the scroll to a parent ScrollContainer
if there is one.

Example: Scrolling a textarea up, it hits the top so it starts scrolling
the window up.

For mouse wheel events, this is done by not marking the event has
handled, so a parent ScrollContainer will process it.

For touch scrolling (motion event), it's complicated by mouse capture.  We used to
release the mouse capture and not mark the event as handled.  But that
had some weird edge cases, especially when scrolling the choices in a
dropdown - since we gave up mouse capture, the touch release would
trigger the dropdown selection.

Now for touch scrolling we keep mouse capture but bubble a new
ScrollPropogate event to parents.