raylib: can now render textures
The root problem turned out to be the order of the triangle indexes.
While dvui was going counter-clockwise for rendering paths, it was going
clockwise for textures.  Now we go counter-clockwise for everything.

SDL and webgl were okay with it, but raylib is more picky, and I'm not
sure why.

The raylib backend is still far from useable (no event processing yet),
but this is good progress.