textSize: don't loop forever if scale is 0
Font.textSize calls Font.textSizeEx in a loop to calculate multiline
text sizes.  textSizeEx in turn calls textSizeRaw to do the calculation
in font units.

textSizeEx was calculating a default max width for the text and
converting it into font units.  But if the scale is 0 (minimized windows
on Windows), then that conversion was going wrong and giving 0, meaning
textSizeRaw would never report that any codepoints had been processed,
leading to an infinite loop in textSize.

Hopefully this fixes https://github.com/david-vanderson/issues/96