FloatingWindow.close(): add warning when no open_flag
FloatingWindowWidget can be passed a *bool as a convenience so that
later code can call floating_window.close() to set that bool to false
and refresh.

Now we warn if close() is called but the widget was never given an
open_flag - in that case close() only does a refresh.