reorderable: test hover rect instead of point
Previously, when testing if a dragged reorderable was ontop of an
existing one, we used only the drag point.  That meant that you could
have most of the dragged reorderable visibly ontop of a slot but since
the drag point was outside the reorder widget it wouldn't work.

Now we use the whole rectangle of the floating reorderable.  The drop
target slot is the first one that intersects the floating reorderable
rect.