reorder: make full widget and more helpers
ReorderWidget wasn't quite a normal widget before, but it was just left
over from when I was getting it to work.

Now it's a normal widget with all the Options, and the layout happens
inside it.

Also add a helper for adding the final slot, and for reordering a slice.