textLayout: overhaul selection changing
To properly implement https://github.com/david-vanderson/dvui/issues/108
the logic needed an upgrade.

TextLayout.Selection gained the Affinity enum.  When the cursor is
between 2 characters that are separated by the end of a line (not a
newline character), affinity says whether the cursor is at the end of
the previous line or at the beginning of the new line.

This allows 3-click line selection and home/end to operate on visual
lines rather than newline chars.  Also left-right in a textEntry can now
position the cursor at the end of a line that doesn't have a newline
character.