textEntry: subtle ctrl-left/right tweak
I never knew, but if the selection is not empty, the pressing ctrl-left/right
moves the cursor to the beginning/end of the selection (and makes it
empty).  On mac option-left/right.

Only if the selection is empty does ctrl-left/right move a word
forward/backward.

I can't find anything that documents this behavior, but I can reproduce
it everywhere, so now we do it too.