beginWait(): add protections against always saying "no event"
beginWait() and waitTime() cooperate to maintain an estimate of the
frame overhead.  It's only adjusted on frames where we wanted to wait a
specific amount of time (like for a timer) AND we didn't get woken up
early for an event.

The second requirement is signalled by the argument to beginWait(), but
when I redid the web backends event handling, I messed it up so it was
always false.

That cause the frame overhead estimate to go wild, which caused heavy
lag.

I added some protections to keep things reasonable even in the face of
accidentally always passing false to beginWait().