floatingWindow: drag-resize from all corners/sides
Bottom right corner is treated specially and happens before widgets
inside the window, which makes it easy to always access.

The rest happen after widgets inside the window, so you can interact
with the widgets right up to the window edge.

Because that makes the hit area very small, I added a 2 pixel margin
around floating windows.  So that space just visually outside the window
can be used for this purpose.