textLayout: better min width
In the example of a chat window, you want textLayout to calculate a min
width even when break_lines is true.  That way a short chat only takes
up as much width as it needs, but a longer one will hit the edge and
wrap.

Now the min width is calculated as if break_lines is false, and the min
height is calculated using the current width.