floatingWindow: fix losing mouse release
With the "adjust from all corners/sides" change, I missed conditioning
the mouse release on having capture.  So with a normal down/up on the
same frame, processEventsBefore would handle the up, then
processEventsAfter would see the down and start capture.  So to the user
it looked like a dropped mouse release.