scroll: log error if child after expanded child
When adding children to scroll areas, it can make sense for the last
child (usually only) added to be expanded.  That way a scrollArea can
have a min size and the textEntry inside can expand to fill the visual
area but scroll if needed.

But it doesn't make sense to add any child to the scrollArea after that.

So add a debug log and outline the added child.

Also add example in the "Debugging and Errors" demo.