textSize: revert back to measuring to advance
In 58d07423831d837a60589a35d519fb9c3a7c628c I thought it made sense to
change textSize to measure to the right edge of the final glyph, but
that was wrong.  A space character has no width, so you have to measure
using the glyph advance.

This resulted in moving the cursor over space characters (in textEntry)
look very weird.

The problem I was trying to solve is caused by floating point
inaccuracies - textSize and renderText can measure the text slightly
differently.

Now we are working around that explicitly.