textLayout: if min width given, don't calculate
Recently textLayout changed so it always reported a min width as if break_lines
was false.

This is very useful for things like chat bubbles where they might be very small
but grow to fill a large space and wrap.

But it made something like a caption hard.  In this case, you want the text to
fit inside the parent.

Now if textLayout is given a min_size_content, it will use that width for its
min width.  If not given, it will calculate a min width as if break_lines was
false.