Options: add max_size_content
* remove scrollArea .expand_to_fit
* remove floatingWindow autoSize() max_size parameter

It's rare to need to specify a maximum size for a widget, but I've run
into enough examples and tried enough alternate solutions that we're
going to try hitting this head on.

I know of 3 cases where this is useful:
* floatingWindow - limit the initial auto sizing
* textLayout - force line breaks at a certain width
* scrollArea - limit how large it gets before starting to scroll

In general it only makes sense if there is something the widget can do
with the content besides show it all (either break lines or scroll).

Previously I tried to use one-off solutions for the specific cases, but
finally realized that during development the solution generally happens at
a higher level than the problem.

The intent is to help in a situation like the Demo > Text Layout where
there is that airplane icon with a caption in the corner.  The caption
text would force that corner widget to be very wide, and this way you
can limit it at the level you think of (the corner box, not down at the
textLayout level).

This is technically not a max size - it is a maximum imposed on the min
size the widget reports.  If a widget has both .max_size_content and
.expand, it could be sized larger than max_size_content.