flexbox: new layout widget and add to Layout demo
Lays out children in left-to-right rows, moving to a new row when space runs
out.  Each row gets at least one child.

The child .expand and .gravity_* are ignored.  It's possible we figure out
meaning for those in the future.

This is an initial implementation, there are many possible enhancements:
* right-to-left rows
* rows filling bottom up
* aligning across rows/cols