Fixed a very heavy memory leak
I never had as much trouble of figuring a memleak out as this one.
Basically, End() is never called all the time, and because of this,
vertex_buffer and index_buffer aren't freed after every frame. Then
overridden, but still with a refcount of 1.
My bad i suppose, but errr... it was actually quite painful to debug.