floatingWindow: don't nudge away from base window
When a floatingWindow first shows, it tries to not be on top of any
existing floatingWindow (so a dialog ontop of another dialog will be
offset a bit).

That code was also triggered for a dialog that was the size of the full
OS window (even if not on top of another dialog).  That's fixed now.

This impacted the demo window on phone-sized screens.