demo: infinite scrolling canvas
@MichaelBelousov was making an application that needed an infinite
scrolling canvas, and it was overly hard to understand what to do.

As one step to fix this, I've added a demo that shows:
* Using scrollArea with a given virtual size
* Adjusting the ScrollInfo to keep a bit of space to scroll in
* Events to pan the area and drag boxes around inside

This is an initial pass, hopefully will improve on it with feedback.

I intend to add at least a zooming feature to show how that could work.