texture target: better start
I ripped back out the "record commands and replay" stuff to make sure I
get the basics right first.  I'll come back and add it again later.

Explicitly track the RenderTarget so that nesting works.  Imagine a
large widget that is caching its appearance that includes a smaller
widget that is also caching.

The offset for rendering and clipping is now applied in the render*
functions rather than separately during RenderCommand replay.