cache: first try at refresh tracking
When CacheWidget is invalidated, it needs to wait some number of frames
for the contents to be rendering properly due to:
* first frame most things are zero sized
* textLayout could have a latency of 2 frames (width, then height)
* alignment could add margin

These things can nest, so there's no way to statically know how many
frames to wait.

Luckily a friend pointed out that widgets call refresh() if they need
another frame, so CacheWidget can wait until there is a frame where no
children call refresh().

This is the first experimental take on trying this.