cache: able to render off-screen widgets
Removed the assumption that all drawing happens inside
dvui.windowRectPixels().  Now the clipping rectangle is the sole source
to understand if a widget is visible or whether drawing will be seen.

Window.begin() sets the clipping rect to windowRectPixels(), but
CacheWidget can set it to a rectangle of a widget that is off-screen, or
outside the scroll viewport.

Note that CacheWidget still doesn't work for web and raylib backends,
because I haven't yet implemented textureCreateTarget()/renderTarget()
for them.