cross-widget dragging
This expands dragging beyond a single widget to allow multiple widgets
to coordinate across a drag.  Each group of widgets needs a unique "drag
name" to know when that kind of drag is happening.

demo: Scroll Canvas now has an example of this.

* dragStart/dragPreStart args changed to include a "drag name"
* new draggingName() function to test if a specific drag is happening
* dragEnd() is no longer called in captureMouse(null), it must be called
  separately

Window.end() checks for an unhandled mouse release event during a drag,
and will end the drag and log a debug message if so.