web: preventDefault on mouse wheel
This allows the scroll canvas demo to use ctrl-wheel without the browser
also doing zooming.

Unsure which events we should preventDefault on, so currently deciding
on a case-by-case basis.

I don't know of a way to "see if dvui handled the event", because that
happens later during the frame render, so we don't know during the
browser event callback if dvui is going to handle the event or not.