font: cleanup line_height_factor
It looked weird that the cursor height included line_height_factor, so
now it doesn't.  That also made textLayout.multiline unnecessary.

Currently we still render selected text background including
line_height_factor, so generally the background color for multiple lines
merges together.  This seems to be what most platforms do, but some of
them only put background color behind each line of text (so not
including line_height_factor).  We might go that way in the future if
there is a compelling reason.

* textEntry cursor is Font.textHeight() (doesn't include
  line_height_factor)
* textLayout: remove unneeded .multiline option