textEntryNumber: improvements when passing a value pointer
@MichaelBelousov reported that textEntryNumber wasn't allowing you to
clear out the text.  This showed that the textEntry inside needed to be
decoupled from the passed value.  Otherwise you can't edit thru an
invalid or empty state to a new valid one.

* Only overwrite the text shown if the passed value is different from
  last frame
* Only write back to the passed value pointer if:
  * enter_pressed or text was changed AND
  * valid parsed value is different from passed pointer

This still leaves room for the displayed value to be different from the
stored value pointer.  In the demo:
* click max on
* enter value "123" (stored 12, 123 is too big)
* unclick max
* stored value is 12, displayed value is 123

This follows the rule that a widget shouldn't change the underlying
state unless the user is interacting with it.  I'm pretty sure that
should override trying to keep the the widget visually in sync.