dialogs: fix bad centering
When a FloatingWindowWidget (that is not given a rect) is first
displayed, it tries to center itself on its "parent" (see below).

@MichaelBelousov reported dialogs launched from a menu were being
centered on the menu. (the menu had to be on the base window, not in a
subwindow)

Previously "parent" meant the previously focused subwindow.  This mostly
worked, but relied on focus not changing, and FloatingWindowWidget had
to delay its own self focus by a frame.

Now "parent" means the last parent subwindow that provided a screen rect
(FloatingWindowWidget yes, FloatingMenuWidget and FloatingWidget no).
So it has more to do with the structure of the code than where focus
happens to be.

It also means dialogs from other threads will center themselves on the
base window rather than whatever subwindow happens to have focus.