Breaking: ContextWidget rework to not contain normal widgets
ContextWidget has always been a little funny in that it wrapped all the
child widgets just to provide right-click support for that area.  But it
had a terrible layout (all children on top of each other).

@MichaelBelousov helped me realize that ContextWidget should not be part
of the normal layout.

Now you pass a screen space pixel rect for where you want right-click to
be active.

ContextWidget still has to be the parent for the extent of the menu code
(to get the .close_popup bubbled event), but now having any normal
widgets inside ContextWidget will log a debug message.

See the example in src/Examples.zig