add textureRead() to get pixels from a texture target
In issue #167 @Deins figured out that in some environments (I can
reproduce in Windows SDL) the texture targets can be reset meaning we
can't rely on them across frames.

So add a function to read the pixels back from a texture target, and
have CacheWidget use that to convert into a normal texture.