snap to pixels: default to true
We used to turn off snap to pixels in Window.init if the total scale of
the screen was >= 2.

But even on a hidpi phone screen, I could see slight flattening of the
sides of text glyphs.

This is caused by having a non-pixel size font so if snap_to_pixels is
false we use a slightly smaller font and enlarge the quad for each
glyph.  I thought that would be okay on a hidpi screen, but either I'm
wrong or there's still some bug in the code.

So for now we'll always default to snap_to_pixels being true, and will
revisit it in the future.  Apps can always turn it off if they need to.