show calculator operands and bump to f64
The current logic would not show the operand you were typing after you pressed
an operator. This makes it so that we show that.

This also fixes an interesting behavior where if you pressed "{operator} ="
then we would perform the operation on the value with an operand of "0". This
means that pressing "8 / =" you would get "inf". Now, pressing the "=" sign is
equivalent to cancelling the pending operation.

Finally, we now use `f64` instead of `f32` to improve precision.