render newline indicators in text selection
Both Chrome and Monaco (VS Code's editor) render newlines idicators like
this at the end of the line. We don't match their behavior completely
because you cannot yet drag a selection to the right to include a
newline without dragging to the next line. However, this can be added
later.

On the other hand, Firefox does not render newline indicators in
<textarea> elements so this is slightly opinionated. However, rendering
the newlines is useful when interacting with multiline so I think that
this is the way to go.