fix touch menus/dropdowns
Reported by @0biWanKenobi

MenuItemWidgets are the normal entries for both menus and dropdowns.
They did not capture the mouse (like most widgets) on mouse/touch down
to support the mouse-specific interaction:
* mouse down to open a dropdown/menu
* mouse motion to select an entry
* mouse up to activate the entry

Separately, scroll will capture the mouse/touch if it sees any touch
motion event.  This supports the normal scroll interaction:
* touch down on a button (button is incidental)
* touch motion enough to start a scroll
* touch scroll (button is not activated)

So what happened was:
* touch down on menu item (didn't capture)
* touch motion went to scroll, which captured it
* touch release went to scroll due to capture

The reason it sometimes worked was if you managed to touch down/release
without a touch motion between, which is quite hard to do.

Solution is for MenuItemWidget to capture on touch down like buttons,
but not on mouse down to retain the mouse interaction.