border: don't force background if uniform
Our method of drawing a border is very simple:
* draw a large rect with border color
* draw a smaller rect with background color

But this doesn't work if there is no background.  We used to force
background on if there was a border.

This commit adds an exception - if the border is uniform in size and
there is no background, render the border as a stroked path.

If the border is non-uniform, we still force background.

There is a small visual difference between the two ways of rendering a
border, but it is almost indistinguishable when the border is thin.

This enables the common pattern of putting a thin uniform border around a
box (boxes usually don't have a background).