windows: initial multi-window support
The first problem to solve for mulitple windows is the application needs
some way to coordinate with the backend on when to register the window
class. A window class of the same name can onl be registered once.
Since the window class is a user-visible detail, we allow the
application to set a custom class, but then this also means they need
to understand when/if that class needs to be registered.

The path I've chosen here is to default to always registering the window
class, which means by default DVUI will crash once a second window
is created if that second window has the same class name as the first.
However, there is a new InitOption `register_window_class` that the
application can set to false to suppress this subsequent call to
RegisterClass. Note that this wouldn't always be set to false though,
if an application is creating multiple kinds of windows with different
classes then they wouldn't want to set this to false. DVUI could try
to track/manage this state via some global mapping of class names to
window registration counts/etc, but...I'd consider this overkill for
what is normally a very simple concept in normal win32 applications.

Note the alternative here would be to take a more "hands off" approach
and just have the application register the class themselves. This is what
I would recommend but what I've included in this PR feels like a more
similar approach to what's already being done in DVUI.